Now the board wont finish the game until the player wins

+++ src/main/java/ar/fiuba/tdd/tp1/rule/AccumulatorRule.java
@@ -30,14 +30,13 @@ public class AccumulatorRule extends BaseRule {
     }
 
     public boolean check() {
-        double accumulator = 0;
-        double cellValueAsInteger;
+        Integer accumulator = 0;
+        Integer cellValueAsInteger;
         boolean atLeasOneCellIsEmpty = false;
 
         for (String cellAsString : cellsAsString) {
             for (Cell cellValue : calculateCellList(cellAsString)) {
                 atLeasOneCellIsEmpty = atLeasOneCellIsEmpty | cellValue.isEmpty();
-
                 cellValueAsInteger = Integer.parseInt(cellValue.getData());
                 accumulator = arithmeticalOperator.apply(accumulator, cellValueAsInteger);
             }
